move common examples build logic into crate
