+++ crates/re_examples/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "re_build_examples"
+name = "re_examples"
 authors.workspace = true
-description = "Build rerun example RRD files"
+description = "Utils for rerun example builds"
 edition.workspace = true
 homepage.workspace = true
 include.workspace = true
@@ -15,14 +15,10 @@ version.workspace = true
 [package.metadata.docs.rs]
 all-features = true
 
-
 [dependencies]
 re_build_tools.workspace = true
-re_examples.workspace = true
 
 # External
 anyhow.workspace = true
-indicatif.workspace = true
-rayon.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_yaml.workspace = true
